+++ NEWS.md
@@ -1,7 +1,10 @@
-# mrgsim.sa (development version)
+# mrgsim.sa 0.2.0
 
-- `sens_plot()` will now plot the all items in `dv_name` by default 
-  (#7). 
+- `sens_plot()` for `sens_grid` output will now plot multiple dependent 
+  variables with a single call (#12).
+
+- `sens_plot()` for `sens_each` output will now plot the all items in 
+  `dv_name` by default (#7). 
 
 - Add `layout` argument to `sens_plot()` to control how plots appear
   when multiple dependent variables are getting plotted (#7).